crack_password_and_username
I tried to input username and password and put it into the firewall.
Then, the data was encrypted by hash function to make it complicated that is really hard to decode. 
So on, I compare each hash function to consider the natural of hash function. 
In other types, I was using some reverse engineering techniques to create a binary file and to find the password.